После сдачи практики :D

+++ app/src/main/java/com/conditer/conditercompany/databasese/priceTable.java
@@ -8,7 +8,7 @@ public class priceTable {
     @PrimaryKey(autoGenerate = true)
     public long id_price = 0;
 
-    public short otdelPrice;
+    public int otdelPrice;
 
     public String namePrice;
